Initial checkin of support for OVPMS-183 Scheduling Workspace
git-svn-id: svn://svn.openvpms.org/openvpms/vpms/trunk@1267 9e4fa352-aafb-0310-9071-df624684024b

+++ src/java/org/openvpms/web/servlet/DownloadServlet.java
@@ -111,7 +111,7 @@ public class DownloadServlet extends HttpServlet {
                                    "inline; filename=\"" + doc.getName()
                                            + "\"");
                 response.setContentType(doc.getMimeType());
-                response.setContentLength((int) doc.getDocSize());
+                response.setContentLength(doc.getDocSize());
                 response.getOutputStream().write(doc.getContents());
 
                 if (_tempDocs.remove(ref)) {
